// src/shared/api/generated/types/NotificationBody.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { NotificationInstanceData } from './NotificationInstanceData'
import type { NotificationMessageData } from './NotificationMessageData'
import type { NotificationSenderData } from './NotificationSenderData'

export type NotificationBody = {
    typeWebhook: string;
    instanceData?: NotificationInstanceData;
    timestamp?: number;
    idMessage?: string;
    senderData?: NotificationSenderData;
    messageData?: NotificationMessageData;
};